ai-content-generator/requirements.txt
langchain
langchain-ollama
langchain-core
datasets
python-dotenv